Define enum in the extra ExecAPI spec
