+++ airflow-core/src/airflow/api_fastapi/execution_api/app.py
@@ -196,6 +196,7 @@ def create_task_execution_api_app() -> FastAPI:
 def get_extra_schemas() -> dict[str, dict]:
     """Get all the extra schemas that are not part of the main FastAPI app."""
     from airflow.api_fastapi.execution_api.datamodels.taskinstance import TaskInstance
+    from airflow.configuration import WeightRule
     from airflow.executors.workloads import BundleInfo
     from airflow.utils.state import TaskInstanceState, TerminalTIState
 
@@ -206,6 +207,7 @@ def get_extra_schemas() -> dict[str, dict]:
         # as that has different payload requirements
         "TerminalTIState": {"type": "string", "enum": list(TerminalTIState)},
         "TaskInstanceState": {"type": "string", "enum": list(TaskInstanceState)},
+        "WeightRule": {"type": "string", "enum": list(WeightRule)},
     }
