+++ task-sdk/src/airflow/sdk/definitions/_internal/abstractoperator.py
@@ -31,12 +31,12 @@ from typing import TYPE_CHECKING, Any, ClassVar, TypeAlias
 import methodtools
 
 from airflow.configuration import conf
+from airflow.sdk.api.datamodels._generated import WeightRule
 from airflow.sdk.definitions._internal.mixins import DependencyMixin
 from airflow.sdk.definitions._internal.node import DAGNode
 from airflow.sdk.definitions._internal.setup_teardown import SetupTeardownContext
 from airflow.sdk.definitions._internal.templater import Templater
 from airflow.sdk.definitions.context import Context
-from airflow.sdk.definitions.weight_rule import WeightRule
 from airflow.utils.trigger_rule import TriggerRule
 
 if TYPE_CHECKING:
